dags/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
dateparser==1.1.1
dnspython==1.16.0
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
idna==3.4
importlib-metadata==4.12.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.1
MarkupSafe==2.1.1
newspaper3k==0.2.8
nltk==3.7
parse==1.19.0
Pillow==9.2.0
pyee==8.2.2
pymongo==3.12.3
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.26.0
requests-file==1.5.1
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.3.1
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
w3lib==2.0.1
websockets==10.3
zipp==3.8.1
